+++ fili-wikipedia-example/src/test/java/com/yahoo/wiki/webservice/application/WikiJerseyTestBinder.java
@@ -6,6 +6,7 @@ import com.yahoo.bard.webservice.application.JerseyTestBinder;
 import com.yahoo.bard.webservice.data.config.dimension.DimensionConfig;
 import com.yahoo.bard.webservice.data.config.metric.MetricLoader;
 import com.yahoo.bard.webservice.data.config.table.TableLoader;
+import com.yahoo.bard.webservice.metadata.TestDataSourceMetadataService;
 import com.yahoo.wiki.webservice.data.config.dimension.WikiDimensions;
 import com.yahoo.wiki.webservice.data.config.metric.WikiMetricLoader;
 import com.yahoo.wiki.webservice.data.config.table.WikiTableLoader;
@@ -48,6 +49,6 @@ public class WikiJerseyTestBinder extends JerseyTestBinder {
 
     @Override
     public TableLoader getTableLoader() {
-        return new WikiTableLoader();
+        return new WikiTableLoader(new TestDataSourceMetadataService());
     }
 }